DRY'ed up MatchBreakdowns
