+++ app/views/breakdowns/MatchBreakdown2017.js
@@ -1,17 +1,13 @@
 import React from 'react';
-import ReactNative from 'react-native';
 import {
   Text,
   View
 } from 'react-native';
 import BreakdownRow from '../../components/BreakdownRow';
-import breakdown from '../../styles/breakdown';
-import images from '../../config/images';
+import breakdownStyle from '../../styles/breakdown';
+import MatchBreakdown from '../breakdowns/MatchBreakdown';
 
-// Override our Image and Text to have specific sizes
-const Image = ({ style, ...props }) => <ReactNative.Image style={[breakdown.imageSize, style]} {...props} />;
-
-export default class MatchBreakdown2017 extends React.Component {
+export default class MatchBreakdown2017 extends MatchBreakdown {
 
   renderRotorView(json_data, rotor_number) {
     var autoEngaged = false
@@ -21,9 +17,9 @@ export default class MatchBreakdown2017 extends React.Component {
     var teleopEngaged = json_data["rotor" + rotor_number + "Engaged"]
 
     if (autoEngaged) {
-      return <Image source={images.checkCircle} />
+      return this.checkCircleImage()
     } else if (teleopEngaged) {
-      return <Image source={images.check} />
+      return this.checkImage()
     }
   }
 
@@ -31,36 +27,18 @@ export default class MatchBreakdown2017 extends React.Component {
     if (value == true) {
       return (
         <View style={{alignItems: 'center', flexDirection: 'row'}}>
-          <Image source={images.check} />
-          <Text style={breakdown.font}>(+ {bonus})</Text>
+          {this.checkImage()}
+          <Text style={breakdownStyle.font}>(+ {bonus})</Text>
         </View>
       );
     } else {
-      return <Image source={images.clear} />
+      return this.xImage()
     }
-  }
-
-  checkImage() {
-    return (
-      <Image source={images.check} />
-    );
-  }
-
-  upArrowImage() {
-    return (
-      <Image source={images.arrows.up} />
-    );
-  }
-
-  downArrowImage() {
-    return (
-      <Image source={images.arrows.down} />
-    );
   }
 
   render() {
     return (
-      <View style={breakdown.container}>
+      <View style={breakdownStyle.container}>
 
         <BreakdownRow data={["Teams", this.props.redTeams, this.props.blueTeams]} vertical={true} subtotal={true} />
