+++ app/views/breakdowns/MatchBreakdown2018.js
@@ -1,33 +1,16 @@
 import React from 'react';
-import ReactNative from 'react-native';
 import {
-  Text,
   View
 } from 'react-native';
 import BreakdownRow from '../../components/BreakdownRow';
-import breakdown from '../../styles/breakdown';
-import images from '../../config/images';
+import breakdownStyle from '../../styles/breakdown';
+import MatchBreakdown from '../breakdowns/MatchBreakdown';
 
-// Override our Image and Text to have specific sizes
-const Image = ({ style, ...props }) => <ReactNative.Image style={[breakdown.imageSize, style]} {...props} />;
-
-export default class MatchBreakdown2018 extends React.Component {
-
-  checkImage() {
-    return (
-      <Image source={images.check} />
-    );
-  }
-
-  xImage() {
-    return (
-      <Image source={images.clear} />
-    );
-  }
+export default class MatchBreakdown2018 extends MatchBreakdown {
 
   render() {
     return (
-      <View style={breakdown.container}>
+      <View style={breakdownStyle.container}>
 
         <BreakdownRow data={["Teams", this.props.redTeams, this.props.blueTeams]} vertical={true} subtotal={true} />
